--- src/components/sequencer/index.ts
@@ -1,1 +1,0 @@
-export * from "./Sequencer";
